refactor: 홈 ui 수정
